Move UsageStatsOptOutState to rest-client
